Drop MAXWALLCMDS to 130, just for now

+++ r_phase9.c
@@ -94,7 +94,7 @@ static void R_UpdateCache(void)
             }
         }
 
-        if ((int8_t)wall->ceilingpicnum == -1) {
+        if (wall->ceilingpicnum == -1) {
             continue;
         }
